--- example/src/Trade.tsx
@@ -1,55 +1,0 @@
-import React from 'react'
-import { useOcean, useTrade } from '@oceanprotocol/react'
-// import { useOcean, usePublish } from '@oceanprotocol/react'
-import { DDO } from '@oceanprotocol/lib'
-import { useState } from 'react'
-import { Metadata } from '@oceanprotocol/lib/dist/node/ddo/interfaces/Metadata'
-
-export function Trade() {
-  const { ocean, accountId } = useOcean()
-  const { buyDT, sellDT, tradeStep, tradeStepText, tradeIsLoading, tradeError} = useTrade()
-  const [did, setDid] = useState<string | undefined>()
-  const handleBuy = async () => {
-    if (!did) { console.error("No DID"); return}
-    const ddo = await ocean.assets.resolve(did)
-    if(ddo){
-      const tx = await buyDT(ddo.dataToken,'1')
-      console.log(tx)
-    }
-    else{
-      console.error("Publish the asset first and create a pricing")
-    }
-  }
-  const handleSell = async () => {
-    if (!did) { console.error("No DID"); return}
-    const ddo = await ocean.assets.resolve(did)
-    if(ddo){
-      const tx = await buyDT(ddo.dataToken,'1')
-      console.log(tx)
-    }
-    else{
-      console.error("Publish the asset first and create a pricing")
-    }
-  }
-  const handleChange = (e: any) => {
-    setDid(e.target.value)
-  }
-  return (
-    <>
-      <div>Trade Datatoken</div>
-      <div>
-        DID <input onChange={handleChange}></input>
-      </div>
-      <div>
-        <button onClick={handleBuy}>Buy 1 DT</button>
-      </div>
-      <div>
-        <button onClick={handleSell}>Sell 1 DT</button>
-      </div>
-      <div>
-        IsLoading: {tradeIsLoading.toString()} || Status: {tradeStepText}
-      </div>
-      
-    </>
-  )
-}
