+++ src/hooks/index.ts
@@ -2,5 +2,4 @@ export * from './useConsume'
 export * from './useMetadata'
 export * from './usePublish'
 export * from './useCompute'
-export * from './useCreatePricing'
-export * from './useTrade'
+export * from './usePricing'
