--- src/hooks/useCreatePricing/index.ts
@@ -1,2 +1,0 @@
-export * from './useCreatePricing'
-export * from './PriceOptions'
