+++ src/hooks/usePricing/README.md
@@ -1,6 +1,6 @@
 # `usePricing`
 
-Hook with helper utilities to create fixed price exchanges or liquidity pools for your data set
+Hook with helper utilities to create fixed price exchanges or liquidity pools for your data set, mint datatokens , buy and sell datatokens
 
 ## Usage
 
@@ -13,7 +13,7 @@ export default function MyComponent() {
   const { accountId } = useOcean()
   const dataTokenAddress = '0x00000'
   // Publish helpers
-  const { createPricing } = useCreatePricing()
+  const { createPricing } = usePricing()
   
   const priceOptions = {
     price: 10,
@@ -27,13 +27,25 @@ export default function MyComponent() {
     await createPricing(dataTokenAddress, priceOptions)
   }
 
+  async function handleMint() {
+    await mint(dataTokenAddress, '1')
+  }
+  async function handleBuyDT() {
+    await buyDT(dataTokenAddress, '1')
+  }
+  async function handleSellDT() {
+    await sellDT(dataTokenAddress, '1')
+  }
   
   return (
     <div>
       <h1>Post for sale</h1>
 
       <p>Your account: {accountId}</p>
+      <button onClick={handleMint}>Mint DT</button>
       <button onClick={handleCreatePricing}>Post for sale</button>
+       <button onClick={handleBuyDT}>Buy DT</button>
+      <button onClick={handleSellDT}>Sell DT</button>
     </div>
   )
 }
