+++ src/hooks/usePublish/DataTokenOptions.ts
@@ -1,5 +1,4 @@
 export interface DataTokenOptions {
-  tokensToMint: number
   cap?: string
   name?: string
   symbol?: string
