--- src/hooks/useTrade/index.ts
@@ -1,1 +1,0 @@
-export * from './useTrade'
